Final Update

+++ src/version04/Server.java
@@ -1,8 +1,8 @@
 package version04;
 
 public class Server {
-	static int nstMin = 0;
-	static int nstMax = 0;
+	static double nstMin = 0;
+	static double nstMax = 0;
 	static double avgTxTime = 5;
 	static double avgRxTime = 5;
 	static int numQueue = 3;
